+++ src/core/Block/Block.tsx
@@ -2,24 +2,40 @@ import React, { Component } from 'react';
 import { default as styled } from 'styled-components';
 import classnames from 'classnames';
 import { SpacingWithoutInsetProp } from '../theme/spacing';
-import {
-  Block as CompBlock,
-  BlockProps as CompBlockProps,
-} from '../../components/Block/Block';
 import { baseStyles } from './Block.baseStyles';
+import { HtmlDiv, HtmlDivProps } from '../../reset';
 
 const baseClassName = 'fi-block';
 
-export interface BlockProps extends CompBlockProps {
+export interface BlockProps extends HtmlDivProps {
   /** Padding from theme */
   padding?: SpacingWithoutInsetProp;
   /** Margin from theme */
   margin?: SpacingWithoutInsetProp;
+  /**
+   * Change block semantics
+   * @default default
+   */
+  variant?: 'default' | 'section' | 'header' | 'nav' | 'main' | 'footer';
+}
+
+class SemanticBlock extends Component<BlockProps> {
+  render() {
+    const { className, variant, ...passProps } = this.props;
+    const ComponentVariant =
+      !variant || variant === 'default' ? HtmlDiv : variant;
+    return (
+      <ComponentVariant
+        {...passProps}
+        className={classnames(baseClassName, className)}
+      />
+    );
+  }
 }
 
 const StyledBlock = styled(
   ({ className, padding, margin, ...passProps }: BlockProps) => (
-    <CompBlock
+    <SemanticBlock
       {...passProps}
       className={classnames(className, {
         [`${baseClassName}--padding-${padding}`]: !!padding,
@@ -36,24 +52,6 @@ const StyledBlock = styled(
  * Used displaying Block with correct styles
  */
 export class Block extends Component<BlockProps> {
-  static section = (props: BlockProps) => (
-    <StyledBlock {...props} variant="section" />
-  );
-
-  static header = (props: BlockProps) => (
-    <StyledBlock {...props} variant="header" />
-  );
-
-  static nav = (props: BlockProps) => <StyledBlock {...props} variant="nav" />;
-
-  static main = (props: BlockProps) => (
-    <StyledBlock {...props} variant="main" />
-  );
-
-  static footer = (props: BlockProps) => (
-    <StyledBlock {...props} variant="footer" />
-  );
-
   render() {
     return <StyledBlock {...this.props} />;
   }
